requirements.txt
# Pinned package versions (generated from active venv pip freeze)
langchain==1.0.3
langchain-openai==1.0.1
openai==2.6.1
python-dotenv==1.2.1
tiktoken==0.12.0
tenacity==9.1.2
requests==2.32.5
